automation/requirements.txt
python-telegram-bot==22.6
psycopg2-binary
APScheduler==3.10.4
google-genai
fastapi==0.115.0
uvicorn==0.30.6
sentry-sdk==2.23.1
python-dotenv
scrapling
telethon
httpx
curl_cffi
patchright
playwright
msgspec
lxml
beautifulsoup4
browserforge
tldextract
